changed merge sort
Merge sort isn't really merge sort anymore. I did this so that the algorithm would only use swaps and therefore not write any duplicates to the main array. In my opinion, it is a worthy sacrifice because the main purpose of this is to visualize the algorithms and not necessaraly do them.